fix windows unit tests

+++ cpp/tests/utility/ProgressBar.cpp
@@ -30,35 +30,34 @@
 #include <thread>
 
 #include "open3d/utility/Parallel.h"
-#include "open3d/utility/ProgressBar.h"
 #include "tests/Tests.h"
 
 namespace open3d {
 namespace tests {
 
 TEST(ProgressBar, ProgressBar) {
-    size_t iterations = 1000;
+    int iterations = 1000;
     utility::ProgressBar progress_bar(iterations, "ProgressBar test: ", true);
 
-    for (size_t i = 0; i < iterations; ++i) {
+    for (int i = 0; i < iterations; ++i) {
         std::this_thread::sleep_for(std::chrono::milliseconds(2));
         ++progress_bar;
     }
-    EXPECT_EQ(iterations, progress_bar.GetCurrentCount());
+    EXPECT_EQ(iterations, static_cast<int>(progress_bar.GetCurrentCount()));
 }
 
 TEST(ProgressBar, OMPProgressBar) {
-    size_t iterations = 1000;
+    int iterations = 1000;
     utility::OMPProgressBar progress_bar(iterations,
                                          "OMPProgressBar test: ", true);
 
 #pragma omp parallel for schedule(static) \
         num_threads(utility::EstimateMaxThreads())
-    for (size_t i = 0; i < iterations; ++i) {
+    for (int i = 0; i < iterations; ++i) {
         std::this_thread::sleep_for(std::chrono::milliseconds(10));
         ++progress_bar;
     }
-    EXPECT_TRUE(progress_bar.GetCurrentCount() >= iterations);
+    EXPECT_TRUE(static_cast<int>(progress_bar.GetCurrentCount()) >= iterations);
 }
 
 }  // namespace tests
